added verbose flag

+++ spotidnldr/clifi.py
@@ -9,9 +9,10 @@ from spotidnldr.tag_embedder import *
 from spotidnldr.converter import *
 
 @click.command()
+@click.option('-v','--verbose', is_flag=True, help="flag for verbosity, usage of this flag gives more output use this for sending or finding errors")
 @click.option('-o', "--output", help="provide a output path for the song explicitly", default="/storage/emulated/0/Songs/")
 @click.option("--url", prompt="enter Url of spotify song", help="The flag which directly sets the url instead of a prompt. if not used the program will prompt you for a url", required=True, type=str)
-def download(url,output):
+def download(url,output,verbose):
     userid=e.SPOTIFY_USER_ID
     clientid=e.SPOTIPY_CLIENT_ID
     clientsecret=e.SPOTIPY_CLIENT_SECRET
@@ -31,12 +32,15 @@ def download(url,output):
         release_date = each["album_data"]["album_release_date"]
         if "/" in name:
             name = name.replace("/\\","" )
+        if "f{name}.mp3" in os.listdir(output):
+            print(f"a file named {name} already in {output} , not downloading it")
+            continue
         if ".temp" not in os.listdir():
             os.mkdir(".temp")
         img_name = dl_jpg(album_cover_url, "./.temp/", name)
         retrived_from_youtube=youtube(key=youtu_key).search(query=name,order="relevance", limit=10, return_indices=3)
         infile = download_yout(url=retrived_from_youtube[0], file_name=name)
-        convert_to_mp3(infile, os.path.join(output,f"{name}.mp3"))
+        convert_to_mp3(infile, os.path.join(output,f"{name}.mp3"), verbose)
         tag_embed(os.path.join(output,f"{name}.mp3"), title=title, artists=song_artists, album=album_name, album_artists=album_artists, release_date=release_date, track_number=track_number, total_tracks=total_tracks, img_path=img_name)
         os.remove(img_name)
         os.remove(infile)
